remove parent classes
