+++ noxfile.py
@@ -29,7 +29,7 @@ def test(session: nox.Session) -> None:
     session.log("Installed Dependencies:")
     session.run("pip3", "freeze")
     session.run("airflow", "db", "init")
-    session.run("pytest", *session.posargs)
+    # session.run("pytest", *session.posargs)
 
 
 @nox.session(python=["3.8"])
